feat: implement LoginBox component for user authentication
